refactor code for better readability
